% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rockercompose.R
\name{make_rockercompose}
\alias{make_rockercompose}
\title{Create a Docker Compose file to build a Rocker image stack}
\usage{
make_rockercompose(
  images_list,
  file_out = "docker-compose.yml",
  org = "rocker",
  dockerfile_path = "dockerfiles"
)
}
\arguments{
\item{images_list}{A list of Rocker image data with variables for assembling
Rocker Dockerfile templates, or a json or yaml file with this data}

\item{file_out}{the filename to write.  If NULL, the YAML is returned as
character data}

\item{org}{the Docker orgaanizzation prefix for images}

\item{dockerfile_path}{The path where dockerfiles are stored, generally
created with \code{\link[=make_rockerfiles]{make_rockerfiles()}}}
}
\value{
the absolute path of file_out, or the Docker Compose YAML
}
\description{
Create a Docker Compose file to build a Rocker image stack
}
