% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rockerfile.R
\name{make_rockerfile}
\alias{make_rockerfile}
\title{Create a Rocker-style Dockerfile from a list of variables}
\usage{
make_rockerfile(
  variables,
  file_out = NULL,
  ref = "master",
  dev_workflow = Sys.getenv("ROCKER_DEV_WORKFLOW", "0"),
  partials_base_url = "https://raw.githubusercontent.com/noamross/rocker-versioned2/"
)
}
\arguments{
\item{variables}{a list of variables for assembling Dockerfiles from Rocker
whisker templates and partialss}

\item{file_out}{the file to write the Dockerfile to. If NULL, returns as
character}

\item{ref}{the git reference of the source repository to use}

\item{dev_workflow}{If "1" builds Dockerfiles using the developer workflow
for faster iteration. If "0", uses the normal workflow that places all
rocker scripts in the base image.}

\item{partials_base_url}{The base URL of the source of partials, e.g.
\verb{"https://raw.githubusercontent.com/user/repo/}}
}
\value{
the absolute path of the Dockerfile, or its contents
}
\description{
Create a Rocker-style Dockerfile from a list of variables
}
