% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rockerfile.R
\name{make_rockerfiles}
\alias{make_rockerfiles}
\title{Create Rocker-style dockerfiles from a list of lists of variables or JSON/YAML data}
\usage{
make_rockerfiles(images_list, out_dir = ".", ...)
}
\arguments{
\item{images_list}{A list of lists of variables for assembling Rocker whisker
templates, or a file or url of JSON/YAML data}

\item{out_dir}{the directory to place the Dockerfiles}

\item{...}{other arguments passed to make_rockerfile}
}
\value{
a vector of Dockerfile file paths
}
\description{
Create Rocker-style dockerfiles from a list of lists of variables or JSON/YAML data
}
